[FE] refactor: Friend 페이지 버튼,모달 분리
